+++ CHANGELOG.md
@@ -58,3 +58,9 @@
 ## 1.0.14
 
 - Make SubcategoryProgressIndicator more flexible.
+
+## 1.0.15
+
+- Make income and expense chart more flexible, inside gap is customizable now.
+- Make subcategory progress indicator customizable, split width by can be set by the user.
+- Make the bottom navigation bar label customizable using the theme data, example of label text size setup is implemented in theme model.